# Vector perf improvements