--- a/vector/materializer.go
+++ b/vector/materializer.go
@@ -100,10 +100,34 @@ func (v *float64s) newMaterializer() materializer {
 	}
 }
 
-func (v *ints) newMaterializer() materializer {
+func (v *int8s) newMaterializer() materializer {
 	var index int
 	return func(builder *zcode.Builder) {
-		builder.Append(zed.EncodeInt(v.values[index]))
+		builder.Append(zed.EncodeInt(int64(v.values[index])))
+		index++
+	}
+}
+
+func (v *int16s) newMaterializer() materializer {
+	var index int
+	return func(builder *zcode.Builder) {
+		builder.Append(zed.EncodeInt(int64(v.values[index])))
+		index++
+	}
+}
+
+func (v *int32s) newMaterializer() materializer {
+	var index int
+	return func(builder *zcode.Builder) {
+		builder.Append(zed.EncodeInt(int64(v.values[index])))
+		index++
+	}
+}
+
+func (v *int64s) newMaterializer() materializer {
+	var index int
+	return func(builder *zcode.Builder) {
+		builder.Append(zed.EncodeInt(int64(v.values[index])))
 		index++
 	}
 }
@@ -149,10 +173,34 @@ func (v *times) newMaterializer() materializer {
 	}
 }
 
-func (v *uints) newMaterializer() materializer {
+func (v *uint8s) newMaterializer() materializer {
 	var index int
 	return func(builder *zcode.Builder) {
-		builder.Append(zed.EncodeUint(v.values[index]))
+		builder.Append(zed.EncodeUint(uint64(v.values[index])))
+		index++
+	}
+}
+
+func (v *uint16s) newMaterializer() materializer {
+	var index int
+	return func(builder *zcode.Builder) {
+		builder.Append(zed.EncodeUint(uint64(v.values[index])))
+		index++
+	}
+}
+
+func (v *uint32s) newMaterializer() materializer {
+	var index int
+	return func(builder *zcode.Builder) {
+		builder.Append(zed.EncodeUint(uint64(v.values[index])))
+		index++
+	}
+}
+
+func (v *uint64s) newMaterializer() materializer {
+	var index int
+	return func(builder *zcode.Builder) {
+		builder.Append(zed.EncodeUint(uint64(v.values[index])))
 		index++
 	}
 }