--- a/vector/types.go
+++ b/vector/types.go
@@ -19,13 +19,19 @@ var _ vector = (*durations)(nil)
 var _ vector = (*float16s)(nil)
 var _ vector = (*float32s)(nil)
 var _ vector = (*float64s)(nil)
-var _ vector = (*ints)(nil)
+var _ vector = (*int8s)(nil)
+var _ vector = (*int16s)(nil)
+var _ vector = (*int32s)(nil)
+var _ vector = (*int64s)(nil)
 var _ vector = (*ips)(nil)
 var _ vector = (*nets)(nil)
 var _ vector = (*strings)(nil)
 var _ vector = (*times)(nil)
 var _ vector = (*types)(nil)
-var _ vector = (*uints)(nil)
+var _ vector = (*uint8s)(nil)
+var _ vector = (*uint16s)(nil)
+var _ vector = (*uint32s)(nil)
+var _ vector = (*uint64s)(nil)
 
 var _ vector = (*arrays)(nil)
 var _ vector = (*constants)(nil)
@@ -64,7 +70,19 @@ type float64s struct {
 	values []float64
 }
 
-type ints struct {
+type int8s struct {
+	values []int8
+}
+
+type int16s struct {
+	values []int16
+}
+
+type int32s struct {
+	values []int32
+}
+
+type int64s struct {
 	values []int64
 }
 
@@ -91,7 +109,19 @@ type types struct {
 	values []zed.Type
 }
 
-type uints struct {
+type uint8s struct {
+	values []uint8
+}
+
+type uint16s struct {
+	values []uint16
+}
+
+type uint32s struct {
+	values []uint32
+}
+
+type uint64s struct {
 	values []uint64
 }